+++ gems/aws-sdk-amplify/lib/aws-sdk-amplify/client.rb
@@ -2871,7 +2871,7 @@ module Aws::Amplify
         params: params,
         config: config)
       context[:gem_name] = 'aws-sdk-amplify'
-      context[:gem_version] = '1.57.0'
+      context[:gem_version] = '1.58.0'
       Seahorse::Client::Request.new(handlers, context)
     end